유저에 필요한 엔티티 및 이넘 작업 중

+++ Spring_Back/src/main/java/com/busanit401/spring_back/domain/PaymentRefund.java
@@ -30,7 +30,7 @@ public class PaymentRefund {
 
     // 해당 월 환불 금액
     @Column(name = "refund_amount", nullable = false)
-    private int refundAmount;
+    private Long refundAmount;
 
     // 토스페이먼츠 환불 처리 후 받는 취소 키
     @Column(name = "cancel_key")
